+++ packages/components/scripts/generateDSRComponents.ts
@@ -659,7 +659,8 @@ import { get${componentName}Css } from '${stylesBundleImportPath}';
           .replace(/this\.props\.disabled = disabled;/, '')
           .replace(/this\.props\.value = state;/, '')
           .replace(/formDisabledCallback\(disabled: boolean\)/, 'formDisabledCallback()')
-          .replace(/formStateRestoreCallback\(state: string\)/, 'formStateRestoreCallback()');
+          .replace(/formStateRestoreCallback\(state: string \| null\)/, 'formStateRestoreCallback()')
+          .replace(/this\.props\.parsedValue/g, 'this.parsedValue');
       } else if (tagName === 'p-input-password') {
         newFileContent = newFileContent
           .replace(/@AttachInternals\(\)/, '')
@@ -679,7 +680,8 @@ import { get${componentName}Css } from '${stylesBundleImportPath}';
           .replace(/this\.props\.disabled = disabled;/, '')
           .replace(/this\.props\.value = state;/, '')
           .replace(/formDisabledCallback\(disabled: boolean\)/, 'formDisabledCallback()')
-          .replace(/formStateRestoreCallback\(state: string\)/, 'formStateRestoreCallback()');
+          .replace(/formStateRestoreCallback\(state: string \| null\)/, 'formStateRestoreCallback()')
+          .replace(/this\.props\.parsedValue/g, 'this.parsedValue');
       } else if (
         tagName === 'p-input-number' ||
         tagName === 'p-input-date' ||
@@ -710,7 +712,8 @@ import { get${componentName}Css } from '${stylesBundleImportPath}';
           .replace(/this\.props\.disabled = disabled;/, '')
           .replace(/this\.props\.value = state;/, '')
           .replace(/formDisabledCallback\(disabled: boolean\)/, 'formDisabledCallback()')
-          .replace(/formStateRestoreCallback\(state: string\)/, 'formStateRestoreCallback()');
+          .replace(/formStateRestoreCallback\(state: string \| null\)/, 'formStateRestoreCallback()')
+          .replace(/this\.props\.parsedValue/g, 'this.parsedValue');
       } else if (tagName === 'p-canvas') {
         newFileContent = newFileContent
           .replace(