+++ packages/components/src/components.d.ts
@@ -982,7 +982,7 @@ export namespace Components {
           * The default date value for the input, in YYYY-MM-DD format (e.g., value='2025-07-02').
           * @default ''
          */
-        "value"?: string;
+        "value"?: string | null;
     }
     interface PInputEmail {
         /**
@@ -1078,7 +1078,7 @@ export namespace Components {
           * The default email address (or comma-separated list of addresses) for the input.
           * @default ''
          */
-        "value"?: string;
+        "value"?: string | null;
     }
     interface PInputMonth {
         /**
@@ -1160,7 +1160,7 @@ export namespace Components {
           * The default month value for the input, in YYYY-MM format (e.g., value='2025-07').
           * @default ''
          */
-        "value"?: string;
+        "value"?: string | null;
     }
     interface PInputNumber {
         /**
@@ -1252,7 +1252,7 @@ export namespace Components {
           * The value of the number input.
           * @default ''
          */
-        "value"?: string;
+        "value"?: string | number | null;
     }
     interface PInputPassword {
         /**
@@ -1339,7 +1339,7 @@ export namespace Components {
           * The password input value.
           * @default ''
          */
-        "value"?: string;
+        "value"?: string | null;
     }
     interface PInputSearch {
         /**
@@ -1435,7 +1435,7 @@ export namespace Components {
           * The search input value.
           * @default ''
          */
-        "value"?: string;
+        "value"?: string | null;
     }
     interface PInputTel {
         /**
@@ -1526,7 +1526,7 @@ export namespace Components {
           * The tel input value.
           * @default ''
          */
-        "value"?: string;
+        "value"?: string | null;
     }
     interface PInputText {
         /**
@@ -1617,7 +1617,7 @@ export namespace Components {
           * The text input value.
           * @default ''
          */
-        "value"?: string;
+        "value"?: string | number | null;
     }
     interface PInputTime {
         /**
@@ -1699,7 +1699,7 @@ export namespace Components {
           * The default time value for the input, in hh:mm or hh:mm:ss format (e.g., value='14:00').
           * @default ''
          */
-        "value"?: string;
+        "value"?: string | null;
     }
     interface PInputUrl {
         /**
@@ -1790,7 +1790,7 @@ export namespace Components {
           * The URL input value.
           * @default ''
          */
-        "value"?: string;
+        "value"?: string | null;
     }
     interface PInputWeek {
         /**
@@ -1872,7 +1872,7 @@ export namespace Components {
           * The default week value for the input, in YYYY-Www format (e.g., value='2025-W27').
           * @default ''
          */
-        "value"?: string;
+        "value"?: string | null;
     }
     interface PLink {
         /**
@@ -3052,7 +3052,7 @@ export namespace Components {
           * The textarea value.
           * @default ''
          */
-        "value"?: string;
+        "value"?: string | null;
         /**
           * Handles wrapping behavior of elements.
           * @default 'soft'
@@ -5167,7 +5167,7 @@ declare namespace LocalJSX {
           * The default date value for the input, in YYYY-MM-DD format (e.g., value='2025-07-02').
           * @default ''
          */
-        "value"?: string;
+        "value"?: string | null;
     }
     interface PInputEmail {
         /**
@@ -5275,7 +5275,7 @@ declare namespace LocalJSX {
           * The default email address (or comma-separated list of addresses) for the input.
           * @default ''
          */
-        "value"?: string;
+        "value"?: string | null;
     }
     interface PInputMonth {
         /**
@@ -5369,7 +5369,7 @@ declare namespace LocalJSX {
           * The default month value for the input, in YYYY-MM format (e.g., value='2025-07').
           * @default ''
          */
-        "value"?: string;
+        "value"?: string | null;
     }
     interface PInputNumber {
         /**
@@ -5473,7 +5473,7 @@ declare namespace LocalJSX {
           * The value of the number input.
           * @default ''
          */
-        "value"?: string;
+        "value"?: string | number | null;
     }
     interface PInputPassword {
         /**
@@ -5572,7 +5572,7 @@ declare namespace LocalJSX {
           * The password input value.
           * @default ''
          */
-        "value"?: string;
+        "value"?: string | null;
     }
     interface PInputSearch {
         /**
@@ -5680,7 +5680,7 @@ declare namespace LocalJSX {
           * The search input value.
           * @default ''
          */
-        "value"?: string;
+        "value"?: string | null;
     }
     interface PInputTel {
         /**
@@ -5783,7 +5783,7 @@ declare namespace LocalJSX {
           * The tel input value.
           * @default ''
          */
-        "value"?: string;
+        "value"?: string | null;
     }
     interface PInputText {
         /**
@@ -5886,7 +5886,7 @@ declare namespace LocalJSX {
           * The text input value.
           * @default ''
          */
-        "value"?: string;
+        "value"?: string | number | null;
     }
     interface PInputTime {
         /**
@@ -5980,7 +5980,7 @@ declare namespace LocalJSX {
           * The default time value for the input, in hh:mm or hh:mm:ss format (e.g., value='14:00').
           * @default ''
          */
-        "value"?: string;
+        "value"?: string | null;
     }
     interface PInputUrl {
         /**
@@ -6083,7 +6083,7 @@ declare namespace LocalJSX {
           * The URL input value.
           * @default ''
          */
-        "value"?: string;
+        "value"?: string | null;
     }
     interface PInputWeek {
         /**
@@ -6177,7 +6177,7 @@ declare namespace LocalJSX {
           * The default week value for the input, in YYYY-Www format (e.g., value='2025-W27').
           * @default ''
          */
-        "value"?: string;
+        "value"?: string | null;
     }
     interface PLink {
         /**
@@ -7469,7 +7469,7 @@ declare namespace LocalJSX {
           * The textarea value.
           * @default ''
          */
-        "value"?: string;
+        "value"?: string | null;
         /**
           * Handles wrapping behavior of elements.
           * @default 'soft'
@@ -7709,7 +7709,7 @@ declare namespace LocalJSX {
         "description": string;
         "compact": boolean;
         "name": string;
-        "value": string;
+        "value": string | null;
         "autoComplete": string;
         "readOnly": boolean;
         "form": string;
@@ -7727,7 +7727,7 @@ declare namespace LocalJSX {
         "description": string;
         "compact": boolean;
         "name": string;
-        "value": string;
+        "value": string | null;
         "autoComplete": string;
         "readOnly": boolean;
         "form": string;
@@ -7750,7 +7750,7 @@ declare namespace LocalJSX {
         "description": string;
         "compact": boolean;
         "name": string;
-        "value": string;
+        "value": string | null;
         "autoComplete": string;
         "readOnly": boolean;
         "form": string;
@@ -7789,7 +7789,7 @@ declare namespace LocalJSX {
         "description": string;
         "compact": boolean;
         "name": string;
-        "value": string;
+        "value": string | null;
         "autoComplete": string;
         "readOnly": boolean;
         "form": string;
@@ -7809,7 +7809,7 @@ declare namespace LocalJSX {
         "description": string;
         "compact": boolean;
         "name": string;
-        "value": string;
+        "value": string | null;
         "autoComplete": string;
         "clear": boolean;
         "indicator": boolean;
@@ -7831,7 +7831,7 @@ declare namespace LocalJSX {
         "description": string;
         "compact": boolean;
         "name": string;
-        "value": string;
+        "value": string | null;
         "autoComplete": string;
         "readOnly": boolean;
         "form": string;
@@ -7874,7 +7874,7 @@ declare namespace LocalJSX {
         "description": string;
         "compact": boolean;
         "name": string;
-        "value": string;
+        "value": string | null;
         "autoComplete": string;
         "readOnly": boolean;
         "form": string;
@@ -7892,7 +7892,7 @@ declare namespace LocalJSX {
         "description": string;
         "compact": boolean;
         "name": string;
-        "value": string;
+        "value": string | null;
         "autoComplete": string;
         "readOnly": boolean;
         "form": string;
@@ -7914,7 +7914,7 @@ declare namespace LocalJSX {
         "description": string;
         "compact": boolean;
         "name": string;
-        "value": string;
+        "value": string | null;
         "autoComplete": string;
         "readOnly": boolean;
         "form": string;
@@ -8205,7 +8205,7 @@ declare namespace LocalJSX {
         "description": string;
         "compact": boolean;
         "name": string;
-        "value": string;
+        "value": string | null;
         "state": TextareaState;
         "message": string;
         "hideLabel": string;