+++ packages/components/CHANGELOG.md
@@ -16,8 +16,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
 
 ### Changed
 
-- `Select`, `Select Option`: extended types of `value` prop to allow `string | number | null`
+- `Input Text`: the global `inputmode` attribute is now passed through to the underlying native `input` element
   ([#4322](https://github.com/porsche-design-system/porsche-design-system/pull/4322))
+- `Input Text`, `Input Number`: extended types of `value` prop to allow `string | number | null`
+  ([#4322](https://github.com/porsche-design-system/porsche-design-system/pull/4322))
+- `Input Date`, `Input Email`, `Input Password`, `Input Search`, `Input Tel`, `Input Time`, `Input Url`, `Textarea`:
+  extended types of `value` prop to allow `string | null`
+  ([#4322](https://github.com/porsche-design-system/porsche-design-system/pull/4322))
+- `Select`, `Select Option`: extended types of `value` prop to allow `string | number | null`
+    ([#4322](https://github.com/porsche-design-system/porsche-design-system/pull/4322))
 - `Multi Select`: extended types of `value` prop to allow `string[] | number[] | null`
   ([#4322](https://github.com/porsche-design-system/porsche-design-system/pull/4322))
 - `Multi Select Option`: extended types of `value` prop to allow `string | number`